Create Binary Variable class skeleton
Idk if this is how I want to do it -- basically, it's
a subclass of clause that really just wraps around a bit.
The idea is that your binary equation will be a tree of nested clauses,
with the roots existing as Binary Variables. If your roots are empty clauses,
the tree is malformed and cannot be analyzed/evaluated -- clauses
do not have any evaluatable value if they have no children.