Fixed issue with number of clauses being generated
Previously, model_parser was generating about 10% too
few clauses; half the issue was due to some bounds issues,
but the other half was more vexing: the "at most" clause
formula in the paper for regions is generated differently
from the others. I made a quick fix which gets us the correct
number of clauses, but I'll need to look into it to make sure
my solution is valid.
If the solver isn't working, look at this first.
Email the writers worst case I guess?